+++ plugin/serve.go
@@ -50,5 +50,5 @@ func Serve(opts *ServeOpts) {
 		}()
 	}
 
-	grpc.NewPluginServer(p.Name, p.SetConnectionConfig, p.SetAllConnectionConfigs, p.GetSchema, p.Execute).Serve()
+	grpc.NewPluginServer(p.Name, p.SetConnectionConfig, p.SetAllConnectionConfigs, p.UpdateConnectionConfigs, p.GetSchema, p.Execute).Serve()
 }